drite-guide/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
sqlalchemy>=2.0,<3.0
asyncpg>=0.30,<1.0
psycopg[binary]>=3.2,<4.0
alembic>=1.16,<2.0
pydantic>=2.11,<3.0
pydantic-settings>=2.9,<3.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
passlib[bcrypt]>=1.7,<2.0
bcrypt==4.0.1
PyJWT>=2.9,<3.0
email-validator>=2.2,<3.0
